Move orbital energy code
