+++ src/qforte/ucc/uccnpqe.py
@@ -339,32 +339,6 @@ class UCCNPQE(UCCPQE):
 
         return residuals
 
-    def build_orb_energies(self):
-        """Calcualtes Hartree-Fock orbital energies to used in determintation
-        of the MP denominators.
-        """
-        self._orb_e = []
-
-        print('\nBuilding single particle energies list:')
-        print('---------------------------------------')
-        qc = qforte.Computer(self._nqb)
-        qc.apply_circuit(build_Uprep(self._ref, 'occupation_list'))
-        E0 = qc.direct_op_exp_val(self._qb_ham)
-
-        for i in range(self._nqb):
-            qc = qforte.Computer(self._nqb)
-            qc.apply_circuit(build_Uprep(self._ref, 'occupation_list'))
-            qc.apply_gate(qforte.gate('X', i, i))
-            Ei = qc.direct_op_exp_val(self._qb_ham)
-
-            if(i<sum(self._ref)):
-                ei = E0 - Ei
-            else:
-                ei = Ei - E0
-
-            print(f'  {i:3}     {ei:+16.12f}')
-            self._orb_e.append(ei)
-
     def initialize_ansatz(self):
         """Adds all operators in the pool to the list of operators in the circuit,
         with amplitude 0.
